+++ app/build.gradle
@@ -6,7 +6,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
 
     compileSdkVersion 23
-    buildToolsVersion "24.0.1"
+    buildToolsVersion "23.0.3"
 
     defaultConfig {
         applicationId "com.github.denisidoro.hellokotlin"
@@ -174,6 +174,7 @@ dependencies {
     androidTestCompile libraries.dexmaker
     androidTestCompile libraries.dexmakerDx
     androidTestCompile libraries.dexmakerMockito
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
     /* Testing
