+++ app/src/main/java/com/github/denisidoro/hellokotlin/core/dagger/InjectorInterface.java
@@ -4,7 +4,7 @@ import android.app.Application;
 
 import com.github.denisidoro.hellokotlin.core.dagger.components.ActivityTopComponent;
 import com.github.denisidoro.hellokotlin.core.dagger.components.ApplicationTopComponent;
-import com.github.denisidoro.hellokotlin.core.pattern.BaseActivity;
+import com.github.denisidoro.hellokotlin.core.pattern.activity.BaseActivity;
 
 public interface InjectorInterface {
