--- app/src/main/java/com/github/denisidoro/hellokotlin/core/pattern/Action.kt
@@ -1,5 +1,0 @@
-package com.github.denisidoro.hellokotlin.core.pattern
-
-interface Action
-
-object START: Action
