--- app/src/main/java/com/github/denisidoro/hellokotlin/core/pattern/AnvilSubscriber.kt
@@ -1,12 +1,0 @@
-package com.github.denisidoro.hellokotlin.core.pattern
-
-import com.beyondeye.reduks.rx.RxStore
-import com.beyondeye.reduks.rx.RxStoreSubscriber
-import trikita.anvil.Anvil
-
-class AnvilSubscriber<S>(store: RxStore<S>) : RxStoreSubscriber<S>(store) {
-
-    override fun onStateChange(state: S?) {
-        Anvil.render()
-    }
-}
