+++ app/src/main/java/com/github/denisidoro/hellokotlin/core/pattern/action/Action.kt
@@ -1,0 +1,5 @@
+package com.github.denisidoro.hellokotlin.core.pattern.action
+
+interface Action
+
+object START: Action
