+++ app/src/main/java/com/github/denisidoro/hellokotlin/counter/CounterActions.kt
@@ -1,6 +1,6 @@
 package com.github.denisidoro.hellokotlin.counter
 
-import com.github.denisidoro.hellokotlin.core.pattern.Action
+import com.github.denisidoro.hellokotlin.core.pattern.action.Action
 
 sealed class CounterActions {
     object INIT : Action
