+++ app/src/main/java/com/github/denisidoro/hellokotlin/counter/CounterActivity.kt
@@ -1,6 +1,6 @@
 package com.github.denisidoro.hellokotlin.counter
 
-import com.github.denisidoro.hellokotlin.core.pattern.BaseActivity
+import com.github.denisidoro.hellokotlin.core.pattern.activity.BaseActivity
 
 class CounterActivity : BaseActivity<CounterState>() {
