+++ app/src/main/java/com/github/denisidoro/hellokotlin/counter/CounterComponent.kt
@@ -1,11 +1,11 @@
 package com.github.denisidoro.hellokotlin.counter
 
-import com.github.denisidoro.hellokotlin.core.pattern.Component
+import com.github.denisidoro.hellokotlin.core.pattern.controller.Component
 
 class CounterComponent(activity: CounterActivity): Component<CounterState>(activity) {
 
     override fun getReducer() = CounterReducer()
-    override fun getInitialState() = CounterState(42)
+    override fun getInitialState() = CounterState(43)
     override val view = CounterView(activity, reduks.store)
 
 }
